Refactor service loader; test on [static dynamic endpoint]
This moves the loader code into `core/base-service`, leaving behind in `services/index.js` only the convenience imports.

Ref #2832
