+++ core/server/server.js
@@ -7,9 +7,9 @@ const Joi = require('joi')
 const Camp = require('camp')
 const makeBadge = require('../../gh-badges/lib/make-badge')
 const GithubConstellation = require('../../services/github/github-constellation')
-const { loadServiceClasses } = require('../../services')
 const { makeBadgeData } = require('../../lib/badge-data')
 const suggest = require('../../lib/suggest')
+const { loadServiceClasses } = require('../base-service/loader')
 const { makeSend } = require('../base-service/legacy-result-sender')
 const {
   handleRequest,
