+++ scripts/export-service-definitions-cli.js
@@ -2,7 +2,7 @@
 
 const yaml = require('js-yaml')
 
-const { collectDefinitions } = require('../services')
+const { collectDefinitions } = require('../core/base-service/loader')
 
 const definitions = collectDefinitions()
